Major update to NWN file translator stuff. Updated API dictionary to remove the journals from the Object type (they do not make sense there). Also fixed a bunch of problems in loading and saving to NWN.

+++ scriptease2/scriptease/translator/io/tools/ScriptEaseFileAccess.java
@@ -393,7 +393,7 @@ public class ScriptEaseFileAccess extends RandomAccessFile {
 	public void writeUnsignedLong(long value, boolean reverseEndianess)
 			throws IOException {
 		byte[] bytes = BitwiseConverter
-				.longToByteArray((long) (value & 0xFFFFFFFFL));
+				.longToByteArray(value & 0xFFFFFFFFL);
 
 		if (reverseEndianess) {
 			bytes = BitwiseConverter.reverseEndian(bytes, INT_BYTE_LENGTH);
